Break redis into it's own project
