+++ Backend/Remora.Discord.Caching/Remora.Discord.Caching.csproj
@@ -23,7 +23,6 @@
 
     <ItemGroup>
       <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="6.0.1" />
-      <PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="6.0.3" />
     </ItemGroup>
 
 </Project>
