+++ Remora.Discord.sln
@@ -89,6 +89,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Remora.Discord.Pagination",
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Interactivity", "Samples\Interactivity\Interactivity.csproj", "{6F2019A8-178E-4E2B-AD0D-B6AA4768BF17}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Remora.Discord.Caching.Redis", "Backend\Remora.Discord.Caching.Redis\Remora.Discord.Caching.Redis.csproj", "{A6767938-4845-4B1A-B6D2-71C8AB5028F9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -199,6 +201,10 @@ Global
 		{6F2019A8-178E-4E2B-AD0D-B6AA4768BF17}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6F2019A8-178E-4E2B-AD0D-B6AA4768BF17}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6F2019A8-178E-4E2B-AD0D-B6AA4768BF17}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A6767938-4845-4B1A-B6D2-71C8AB5028F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A6767938-4845-4B1A-B6D2-71C8AB5028F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A6767938-4845-4B1A-B6D2-71C8AB5028F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A6767938-4845-4B1A-B6D2-71C8AB5028F9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -225,6 +231,7 @@ Global
 		{D1FBE943-B854-456B-9C3B-C49EAB24AE61} = {D1A34922-E02F-492B-AEF5-9B34E81CE293}
 		{9F9B3DFD-891A-4823-9BF7-E58CC52989C5} = {6E2D59EB-648D-4968-8A63-8D8E3813EC86}
 		{6F2019A8-178E-4E2B-AD0D-B6AA4768BF17} = {D1A34922-E02F-492B-AEF5-9B34E81CE293}
+		{A6767938-4845-4B1A-B6D2-71C8AB5028F9} = {26E93608-29BA-45C7-993A-AA24A49D6013}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {02545F40-E06C-4956-A77D-02EB2D41A964}
